Neural Networks/Alzheimer MRI Analysis/requirements.txt
streamlit==1.31.0
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
pandas==2.2.0